requirements.txt
Flask==1.1.2
pytest==5.3.0; python_version > "3.0"
pytest==4.6.6; python_version < "3.0"
gunicorn
google-cloud-vision==1.0.0
google-cloud-storage==1.28.0
Wand==0.5.9
requests
pillow
google-api-python-client
